Some label settings added to vizBar. Background settings panel first draft created (not fully functional and missing some labels perhaps).

+++ VisualizationAPI/src/org/gephi/visualization/api/color/Color.java
@@ -93,7 +93,7 @@ public final class Color {
     }
 
     public Color(java.awt.Color c) {
-        this(c.getRed(), c.getGreen(), c.getBlue(), c.getTransparency());
+        this((byte) c.getRed(), (byte) c.getGreen(), (byte) c.getBlue(), (byte) c.getTransparency());
     }
 
     /*
